Agregadas clases dao e implementando el action validate del login

+++ src/java/clasesDAOinterfaces/UsuarioDAO.java
@@ -12,6 +12,6 @@ import entities.Usuario;
  *
  * @author agustin
  */
-public interface UsuarioDAO extends BaseGenericDAO<Usuario, Long>{
+public interface UsuarioDAO extends BaseGenericDAO<Usuario, Integer>{
     
 }
